Much improved k8s support
A couple of really critical features were missing in the previous k8s commit:

1. No OpenVPN support (added)
2. Proper cleanup of tertiary resources was missing (fixed)
3. No support for private docker registries (partially working)

For OpenVPN: We take a completely different approach with using OpenVPN with Kubernetes.
We make exclusive use of the 'helm' k8s package manager, which comes with a ready-to-use
OpenVPN server running directly within the k8s cluster itself.

We connect to it without any need to run OpenVPN inside the containers like we do with
VMs. After the server is running, we setup iptables port forwarding rules that allow
the containers to report their telemetry and logs back to CloudBench over the VPN.
Everything works transparently and is setup for the user by this script.

> Usage:
> $ ./cb -f exit
> $ util/setuphelm.sh

At the end of the script, we have preliminary support to also setup a helm-based
Docker registry (based on Object Storage), which is extremely convenient. This doesn't
fully work yet, however, because it requires the Docker daemon on each k8s node
to be modified to support access to insecure private docker registries. This is not possible
with public "aaS" k8s clusters because we don't have access to the k8s hosts.
Thus, it will only work with pre-existing clusters until we find a workaround.

> Dependencies:
> Install kubectl
> Install helm

Overall, the whole process is totally transparent and seamless. After running the
setup script, a fully-functional VPN is working without any prior dependencies,
provided by the k8s cluster itself. All subsequent CB commands work great, including
all logs and telemetry.