libcloud header logging
We need the ability to log HTTP headers to debug what happens
with libcloud.

For the response HTTP headers, that's not problem => We get that
from the libcloud library itself.

For the request headers, libcloud doesn't make that easy, so we have to
bypass it.

What we're trying to do here is solve a conundrum. Under the covers,
libcloud uses /usr/lib/python2.7/httplib.py as it's primary HTTP
transport library.

httplib, has the ability to debug itself, but instead of using the logging
library like a "good" python library should, it just prints to stdout.
In a separate patch (via Dockerfile) we fix httplib to work correctly
and use a logger. We can't really monkey patch this, and sending a python
patch upstream would take too long

That being said, after it uses the logging library correctly, we still
have to capture the log output the headers ONLY when something bad happens.
We don't want to spew a bunch of header logs to the logging system
for the 99% of the time when things are working fine.

To solve that problem, we trap the (newly fixed) http header log messages
from the httplib library into a StringIO buffer. Then, if we happen
to hit a python exception in libcloud, we dump those headers to the log
if and only if there is an exception.

That's what all the mess below is for.

So, below, if we don't carry the patch to httplib, we only log the response
headers (which are usually what matters). If we carry the httplib patch
(sold separately), then we will also log the request headers.