Fully deterministic round-robin scheduling
We've had round-robin provisionig for a while, but it was not deterministic,
meaning that if you reset CB, the next time you create a batch of VMs,
there was no gaurantee the same VMs would go to the same places.

The commit makes it so that VMCs are sorted and then assigned VMs to them
in such a way that the head-VM of the the AI always goes first followed by
the remaining VMs in the AI in the same order in which they were assigned.

This is really important when you have a cluster of heterogeneous hypervisors
that have different sizes or different performance characteristics and
you want things to perform in the same way each time.