Incorporate a semi-real scheduler.
This change modifies the existing round-robin scheduling to make sure
that we keep all VMCs "equally full" and not simply just do round-robin
scheduling.

We do this by adding an additional counter at the VMC level. This is necessary
because access to the counter for scheduling has to be synchronized by
all VM attach operations so that the counter is increased only when a VMC
is actually being selected. This counter works as a 3rd type of admission
control, similar to regular admission control and speculative admission
control, but specifically for the number of VMs per VMC. The other admission
control counters only restrict the total "global" number of VMs or AIs
and so forth, so we add a 3rd one to do per-VMC capacity scheduling.

This change allows us to keep all VMCs full despite the errors our
conditions where we target specific VMCs with VMs.