broken readline history file function
In newer version of python 2.7 (presumably v3 as well) when
combined with Ubuntu 18.04, the `readline.write_history_file` function
has changed to actually use the `rename` system call instead of
updating the history in place.

This causes problems in Docker-based environments, because we attempt
to preserve the CB history file by including it via `docker -v` as
a bind mount. Since python is unable to replace the file (Because
the container doesn't own it), this fails.

The solution is to use a temporary file and move the contents of the
temporary file to the bind-mounted file.

This solution should work with and without docker as well and be transparent.