This commit greatly improves the debugging of the Load Manager.
From now on, when `appnoload`/`dont_start_load_manager` is set, we
actually execute the load manager, but in "debug" mode. This means that
all load manager startup scripts are created on the correct instance(s),
but they are not actually executed. This change makes the current code
compatible with the updated documentation in
https://github.com/ibmcb/cbtool/wiki/HOWTO:-Debug-initial-setup, and
especially https://github.com/ibmcb/cbtool/wiki/FAQ-U#uq19.