node selector support
K8S has a lot of scheduling options. This commit further extends use of those
choices to use "nodeSelectors" so that CB can issue new pods without having
to actually choose a host while the hosts that we care about are grouped together
on the k8s-side of the house instead of specifying them explicitly in our configuration.