Add ability to deal with multi-vnic instances
A large portion of CloudBench-supported cloud providers allow for multiple
vnics in a single instance. With this commit, CB can deal with it, not
only by automatically detecting and runnning dhclient on multiple
additional vNICs, but by also collecting all the IP addresses and
storing those on the VM-specific variable "extra_cloud_ips".