Incorporate a basic destroy-time backoff mechanism
A common error-handling issue with public clouds is detaching VMs.

Currently, we have a mechanism already (that we depend on a lot, actually)
that will attempt to repetively detach resources forever until it succeeds.
This is really important when we are spending money on a public cloud
running a benchmark.

This commit allows this mechanism to backoff during this waiting period.
Rather than re-submit the same detach request every 5 seconds, it provides
a very simple backoff mechanism that you can configure a maximum for,
so that we don't inundate the cloud with too many operations to perform,
particularly if the cloud is transiently unavailable.