Don't use the public network if a load balancer is not enabled
Previously (for the wrk2 workload), we introduced a change that
allowed ingress haproxy traffic to come in over the public network, while
allowing egress haproxy traffic to exit over the private network to
mimic our customers.

Don't do this if there is no haproxy in the picture. =)

The problem with doing this is that the firewall will block this traffic
if we really wanted.

As a result, I have also included a patch that allows access on the public
network for the firewall only from other dropls (just like the private network).