Retry VM lookups
Looking up the "status" of a VM is by far the most high-frequency
API call we make to various cloud providers. Because we make it so frequently,
it is prone to failure. If any of these calls fail, it causes the VM creation
to abort prematurely, even in the face of intermittent network activity.

For libcloud-based clouds (at least), let's retry the call, which by default
will re-issue the call every 1-second within a 5-second period until the call
succeeeds.