Opt-in to extended [host]names
"extended" means that we will append the AI name and VM roles
to the end of the VM's name for rsyslog purposes. This is rather
convenient because it translates to the VM's hostname which makes
log messages easier to lookup when you know the AI and role information.

Let's default this functionality to false, because it can unpredictably reach the Linux
64-byte hostname limit. The remainer of the hostname is under the control of the user,
because they can pick what the "name" of the cloud should be.

If you want the additional information available to you in rsyslog messages,
then you can "opt-in" here by setting it to $True