Nested improvements
1. This PR speeds up container startup a little bit in the VM by using
better rsync options when the container is deployed.
2. We also provide initial support for hugepages

We spent some time testing hugepages in CloudBench.

Here are a handful of changes that help prepare to enable it in a nested
environment (particularly with the PLM adapter).

NOTE: We don't actually configure the guest kernel here yet in
CloudBench to choose the number of hugepages as that is HIGHLY dependent
on the application, but we do put the limits in place (and an example
option in MySQL).

For it to actually take effect, one would either need to reserve the
pages or turn on THP at the guest level via CloudBench in a followup PR.