`platform.linux_distribution()` Deprecated and Removed as of Python3.7
During some testing of `cbtool` on Python3.8, ran into an issue with the
dependencies checker. As it turns out the `linux_distribution()` method
was removed from the `platform` module in the standard lib in Python3.7.

* Try to import `distro`. If version below 3.7, pass, else, raise
exception. `distro` is a requirement for Python >= 3.7
* Try to use `platform.linux_distribution()`. If an `AttributeError` is
raised (indicating the `linux_distribution` method is no longer
available), use `distro.name()`, `distro.version()`, `distro.id()` to
obtain the same information from the `distro` package.